Split out tests.
